requirements.txt
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.3
attrs==25.4.0
backoff==2.2.1
boltons==21.0.0
bracex==2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
cryptography==46.0.3
dill==0.4.0
distro==1.9.0
exceptiongroup==1.2.2
face==24.0.0
faiss-cpu==1.13.2
gitdb==4.0.12
gitpython==3.1.46
glom==22.1.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.2.0
langchain-core==1.2.6
langchain-google-genai==3.0.1
langfuse==3.11.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
loguru==0.7.3
markdown-it-py==4.0.0
mccabe==0.7.0
mcp==1.23.3
mdurl==0.1.2
numpy==2.4.0
openai==2.14.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pathspec==1.0.1
peewee==3.18.3
platformdirs==4.5.1
pluggy==1.6.0
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pylint==4.0.4
pytest==9.0.2
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.30.0
ruamel-yaml==0.19.1
ruamel-yaml-clib==0.2.14
ruff==0.14.10
semgrep==1.146.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.1.2
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tomli==2.0.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uuid-utils==0.12.0
uvicorn==0.40.0
wcmatch==8.5.2
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
